requirements.txt
ray[all]==2.7.0
snowflake-connector-python[pandas,secure-local-storage]==3.2.0
python-dotenv==1.0.0
pandas
cryptography==41.0.4
pyarrow